Chien2

+++ ModerationService.API/Fearture/Command/Moderations/RejectCourseCommand.cs
@@ -41,7 +41,7 @@ namespace ModerationService.API.Fearture.Command.Moderations
                     IsSeen = false,
                     NotificationContent = request.ReasonWhyReject,
                     SendDate = DateTime.Now,
-                    Course_Id = moder.CourseId,
+                    Course_Id = (int)moder.CourseId,
                 };
                 await _publish.Publish(notificationForAdminBussiness);
